+++ CHANGELOG.rst
@@ -26,6 +26,7 @@ Change Log
 - Administrators can now view pages as if they were Assistants; allowing them to (for example) use the data entry forms that enforce double-checking without needed to create a separate account
 - Fixed χ² test in motion statistics, and refactored and moved motion statistics page
 - Teams, like adjudicators, no longer need to have an institution
+- Added page allowing update of adjudicator scores in bulk
 - Made speaker standings more configurable
     - Second-order metrics can now be specified
     - Added trimmed mean (also known as high-low drop)
@@ -42,6 +43,7 @@ Change Log
     - Current team standings page now shows silent round results if "Release all round results to public" is set
     - The Consensus vs Voting options for how ballots work has not been split into two settings: one for preliminary rounds and one for elimination rounds
     - Speaker scores now show as integers (without decimals) where the tournament format would not allow decimals.
+    - Added a page showing a list of institutions in the tournament
 - Switched to using a Websockets/Channels based infrastructure to better allow for asynchronous updates. This should also ameliorate cases where the memcachier plugin expired due to inactivity which would then crash a site. Notes for those upgrading:
     - On Heroku: You should remove the memcachier plugin and instead add 'heroku-redis' to any instances being upgraded
     - Locally: You should recreate your local_settings.py from the example file